% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPV.R
\name{NPV}
\alias{NPV}
\title{Calculate total NPV}
\usage{
NPV(x, y)
}
\arguments{
\item{x}{Numeric vector or list}

\item{y}{Numeric vector}
}
\value{
A vector containing the sum of the NPV
}
\description{
Calculate total NPV
}
\examples{
NPV(1:10, 0.05)

}
