% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPV.R
\name{NPV}
\alias{NPV}
\title{Calculate total NPV}
\usage{
NPV(x, y)
}
\arguments{
\item{x}{Numeric vector or list}

\item{y}{Numeric vector}
}
\value{
A vector containing the sum of the NPV
}
\description{
Calculate the total NPV from yearly profits and a discount rate
}
\examples{
NPV(1:10, 0.05)
NPV(c(200, 180, 240), 0.08)
}
